refactor: scale을 자유롭게 설정 가능하도록 개선

+++ src/main/kotlin/com/danvhae/oscar/console/badapple/processors/VideoProcessor.kt
@@ -6,7 +6,7 @@ import org.bytedeco.javacv.OpenCVFrameConverter
 
 object VideoProcessor {
     fun FFmpegFrameGrabber.nextFrameToBraille(
-        scale: Int
+        scale: Double
     ): String{
         return OpenCVFrameConverter.ToMat().use { converter ->
             val frame = grabImage() ?: throw NoSuchElementException("no more frame")
